feat: comprehensive stack improvements - CI/CD, build system, code quality
- Enhanced CI/CD workflows with multi-platform testing (Ubuntu, Windows, macOS)
- Added security scanning (CodeQL), performance benchmarking, and docs validation
- Improved build system with enhanced Makefile (parallel builds, colored output)
- Added code quality tools (Prettier, EditorConfig)
- Enhanced TypeScript configuration with strict type checking (balanced)
- Added optimization scripts for production builds
- Updated Dependabot configuration and release workflow (ARM64 support)
- Created comprehensive documentation (CHANGELOG, CONTRIBUTING)
- Fixed cross-platform compatibility issues
- Fixed critical bugs: bitlist padding overflow, offset calculation overflow, container offset integer overflow
- All tests passing (59/59)

+++ c-skel/src/ssz_stream.c
@@ -180,8 +180,11 @@ int ssz_stream_root_from_buffer(
           return SSZ_ERR_NON_CANONICAL;
         }
         
-        uint32_t field_offset = bytes[offset] | (bytes[offset+1] << 8) | 
-                               (bytes[offset+2] << 16) | (bytes[offset+3] << 24);
+        /* Fix: ensure safe cast to unsigned - prevent sign extension */
+        uint32_t field_offset = ((uint32_t)bytes[offset]) | 
+                               ((uint32_t)bytes[offset+1] << 8) | 
+                               ((uint32_t)bytes[offset+2] << 16) | 
+                               ((uint32_t)bytes[offset+3] << 24);
         offset += 4;
         
         /* Validate offset points into data region */
